move some class to another package
