+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar">
         <activity android:name=".features.player.PlayerDetailActivity"></activity>
-        <activity android:name=".features.favorite.MatchActivity" />
+        <activity android:name=".features.favorite.FavoriteActivity" />
         <activity android:name=".features.team.TeamActivity" />
         <activity
             android:name=".features.match.search.SearchMatchActivity"
